mandatory part complete without check

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: acampo-p <acampo-p@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/03 15:01:44 by acampo-p          #+#    #+#             */
-/*   Updated: 2022/12/03 16:49:43 by acampo-p         ###   ########.fr       */
+/*   Updated: 2022/12/05 16:05:17 by acampo-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (src[index] != '\0' && index < size - 1)
